Use cubic instead of linear interpolation in predict
The cubic interpolation is what is suggested in Cleveland and
Grosee (1991) and also what R uses. It makes the prediction function
once differentiable which is what I think most people would
expect from LOESS.

Also, fix bug in use of partialsort!. Only the q'th element was
ensured to be at the right localtion instead of all the first q
elements.

Use evalpoly from Julia and bump minimum requirement to Julia 1.6