Change split point calculation in KD-tree construction
This tries to mimic the splitting of the original Loess implementation
which is used by R. The implementation is based on reverse enginerring
of the behavior as the rules are only loosely described in the original
paper. With the rules described in the comment we are able to match the
splits of R.

When adding tests, I realized that the weight calculation in the local
regression were off by a square root. They were computed as the diagonal
Elements of W in `inv(X'*W*X)*X'*W*y` but we applied them to X and y before
computing the OLS estimates so the weights were squared.

The signatures have also been loosened to allow more element types. This
made it easier to test with the cars dataset from R.

I've added a lot of `@debug` statements to made it easier to follow the
KD-tree construction.